# Python: Speedup with cython extension