--- a/python/MANIFEST.in
+++ b/python/MANIFEST.in
@@ -1,0 +1,2 @@
+# include the c extensions
+recursive-include flatbuffers *.h *.pxd *.pyx