--- a/python/flatbuffers/number_types.py
+++ b/python/flatbuffers/number_types.py
@@ -12,10 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import ctypes
-import collections
 import struct
-from ctypes import sizeof
 
 from . import packer