… python/flatbuffers/fastcodec.pyx
… src/idl_gen_python.cpp
… tests/MyGame/Example/Monster.py
… tests/MyGame/Example/TestSimpleTableWithEnum.py
… tests/MyGame/Example/Vec3.py
… tests/py_test.py