Tested in `python2.6.9`, `python2.7.10`, `python3.5.0`, `PyPy 2.6.1(Python 2.7.10)`.

```shell
$ cd flatbuffers/python
$ python2.7 setup.py build  # build package in-place
$ mv build/lib.macosx-10.11-intel-2.7/flatbuffers ../tests/  # move package to tests directory
$ cd ../tests
$ python2.7 py_test.py 10000 10000 10000  # import flatbuffers locally
----------------------------------------------------------------------
Ran 58 tests in 0.007s

OK
vtable deduplication rate: 433385.41/sec
traversed 10000 288-byte flatbuffers in 0.31sec: 31841.75/sec, 8.75MB/sec
built 10000 288-byte flatbuffers in 0.13sec: 75267.23/sec, 20.67MB/sec
$ rm flatbuffers/fastcodec.so  # remove extension, test in pure python
$ python2.7 py_test.py 10000 10000 10000
----------------------------------------------------------------------
Ran 58 tests in 0.066s

OK
vtable deduplication rate: 12294.36/sec
traversed 10000 288-byte flatbuffers in 1.92sec: 5212.67/sec, 2.55MB/sec
built 10000 288-byte flatbuffers in 4.76sec: 2099.42/sec, 0.58MB/sec
```