@layzerar 2015-10-25
https://github.com/google/flatbuffers/issues/284