@rw 2015-10-25
Cool! It doesn't appear to test Cython _and_ Python. How do I do that?