@layzerar 2015-10-25
Force pushed! Add `MANIFEST.in` file, so that the `*.h` and '*.pyx' will be include into the source distribution.