@layzerar 2015-10-25
@rw `import flatbuffers`, if `flatbuffers.fastcodec`  is not `None` means that you using cython version currently.