@rw 2015-10-25
@layzerar How can I choose the plain-Python code when I have Cython installed?