@layzerar 2015-10-26
@rw  Yes, It passes all the tests.

In the python version of `Builder`, `self.Bytes` means the whole buffer, `self.Head()` means the offset of the payload, so `self.Output()` is equal to `self.Bytes[self.Head():]`.

The cython version is based on the `C++` implement, access the whole buffer is not a good idea. So `self.Bytes` is always equal to `self.Output()`, and `self.Head()` always return `0`.