@layzerar 2015-10-26
You can try to remove `fastcodec.so` file in the package or just use `from flatbuffers.builder import Builder` instead of `from flatbuffers import Builder`, then it will fallback pure-python version.

Switching between cython and pure-python is very easy to implement. Can I make a pull request after this pull request? This pull request contains too much.