@layzerar 2015-10-29
Update.

1. Rebase master.
2. We can switch between two implementations by `flatbuffers.force_implementation('python')` and `flatbuffers.force_implementation('cython')`.
3. When Cython extension building to fail, fallback to plain Python version.