@rw 2015-10-30
@layzerar I got it to test locally. This looks interesting. I like seeing such a speedup!

A few things I'd like to see:

1) The diff is confusing and seems to have extra changes (like whitespace modifications in py_test.py). Could you work on cleaning up the diff?

2) We need to be able to run CPython and Cython tests from the same script. Would you look into modifying PythonTest.sh and py_test.py so that it happens?

Keep up the good work!